+++ CHANGELOG.md
@@ -14,7 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Locales: typos in `almanax` command's help
+- Command `almanax`: typo in help's translations
+- Command `about`: typo in translations
 
 ## [0.2.4] - 2021-02-23