+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Command `almanax`: typo in help's translations
 - Command `about`: typo in translations
+- Command `issue`: missing command's name in help's examples
 
 ## [0.2.4] - 2021-02-23