+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.5] - 2021-02-27
+
 ### Added
 
 - Argument `time`: parse hours, minutes and seconds from string
@@ -191,7 +193,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Task `almanax`: automatically fetch the [almanax](http://www.krosmoz.com/en/almanax) of the day at midnight
 - Task `twitter`: fetch news from twitter in (almost) real-time
 
-[unreleased]: https://github.com/brinkflew/krosmobot/compare/v0.2.4...HEAD
+[unreleased]: https://github.com/brinkflew/krosmobot/compare/v0.2.5...HEAD
+[0.2.3]: https://github.com/brinkflew/krosmobot/compare/v0.2.4...v0.2.5
 [0.2.3]: https://github.com/brinkflew/krosmobot/compare/v0.2.3...v0.2.4
 [0.2.3]: https://github.com/brinkflew/krosmobot/compare/v0.2.2...v0.2.3
 [0.2.2]: https://github.com/brinkflew/krosmobot/compare/v0.2.1...v0.2.2